Modified Page Dialog to clearly show image information in relation to a page

+++ src/main/java/com/meritoki/app/desktop/retina/model/document/Page.java
@@ -210,7 +210,7 @@ public class Page {
 	public BufferedImage getShapeBufferedImage(Shape shape) {
 		BufferedImage bufferedImage = null;
 		if (this.getBufferedImage() != null) {
-			if(shape.position.point.x >= 0 && shape.position.point.y >= 0 && shape.position.dimension.height > 0 && shape.position.dimension.width > 0) {
+			if(shape.position.point.x >= 0 && shape.position.point.y >= 0 && (int)shape.position.dimension.height > 0 && (int)shape.position.dimension.width > 0) {
 				bufferedImage = this.getBufferedImage().getSubimage((int)shape.position.point.x, (int)shape.position.point.y, (int)shape.position.dimension.width, (int)shape.position.dimension.height);
 			} else {
 				logger.error("MAJOR ERROR THAT NEEDS TO BE FIXED shape.position="+shape.position);
